requirements.txt
beautifulsoup4>=4.12.0
playwright>=1.40.0
requests>=2.31.0
aiohttp>=3.9.0
openai>=1.6.0
tavily-python>=0.3.0
anthropic>=0.8.0
fasttext-wheel>=0.9.2
langdetect>=1.0.9
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
datasketch>=1.6.4
detoxify>=0.5.2
spacy>=3.7.0
pandas>=2.1.0
pyarrow>=14.0.0
datasets>=2.15.0
httpx>=0.25.0
tenacity>=8.2.3
redis>=5.0.0
pyyaml>=6.0.1
click>=8.1.7
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
tqdm>=4.66.0
loguru>=0.7.2
prometheus-client>=0.19.0
streamlit>=1.29.0
plotly>=5.18.0
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.0
